This patch looks bigger than it is... it renamed PotentialCombatTargets to match the style of other variables...
This also enforces not targeting targets until they are in weapons range (if they arent in weapons range combatMissionCtrl should be moving us toward them)
This also adds a few more innerspace commands: ListHighvalueTargets, ListLowValueTargets, ListPostentialCombatTargets
reminder: you can type "help" or "QuestorCommands" or "ListQuestorCommands" at the innerspace console and it'll list the available commands we provide

+++ Questor.Modules/Lookup/Settings.cs
@@ -405,8 +405,8 @@ namespace Questor.Modules.Lookup
 
         public int DoNotSwitchTargetsIfTargetHasMoreThanThisArmorDamagePercentage { get; set; }
 
-        public int DistanceNPCFrigatesShouldBeIgnoredByPrimaryWeapons { get; set; } //also requires SpeedFrigatesShouldBeIgnoredByMainWeapons
-        public int SpeedNPCFrigatesShouldBeIgnoredByPrimaryWeapons { get; set; } //also requires DistanceFrigatesShouldBeIgnoredByMainWeapons
+        public double DistanceNPCFrigatesShouldBeIgnoredByPrimaryWeapons { get; set; } //also requires SpeedFrigatesShouldBeIgnoredByMainWeapons
+        public double SpeedNPCFrigatesShouldBeIgnoredByPrimaryWeapons { get; set; } //also requires DistanceFrigatesShouldBeIgnoredByMainWeapons
         
         public bool AddWarpScramblersToPrimaryWeaponsPriorityTargetList { get; set; }
         public bool AddWebifiersToPrimaryWeaponsPriorityTargetList { get; set; }
